feat(ui): detect lost render batches and recover the WebView app
Blazor's WebViewRenderer requires render batch acknowledgements to arrive in
order. Backgrounding the app on Android stops the WebView renderer consuming
messages while .NET keeps rendering into it, so batches are lost; the first
acknowledgement after the gap makes NotifyRenderCompleted throw, and it never
resyncs - OnAfterRenderAsync stops firing for good, and components that create
their JS objects there are never initialized. Recovery today is incidental:
Blazor's fatal-error overlay becomes visible and a MutationObserver notices.

Recovery moves out of BrowserInit into AppRecovery and gains a direct signal -
a gap in the render batch sequence, seen as it arrives rather than several hops
later through an overlay this app deliberately styles to zero size. The gap size
is logged, which is what separates a missed reload from one that never fired.

RenderSync already owned the host message seam, so the detector observes it
rather than hooking it twice. Its handler list only covered iOS and Tizen
(__dispatchMessageCallback) though: Android keeps its handler on
window.external.__callback, so the seam had never installed on the one platform
this matters on, and renderSyncHook was null there.
